+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ## Version 0.4.1 ()
+* Add get function to ExtensionPoint to get a specific extension registered with the extension point.
+* Add clear function to ExtensionPoint to Empty the point's registry of all extensions.
+* Add unregister function to ExtensionPoint to 
 ### Features
 * Added types. To make use of them in typescript, import pluggable-electron/renderer pluggable-electron /main or pluggable-electron/preload in the respective parts of your application.